requirements.txt
asgiref==3.3.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.8.6
pytz==2021.3
sqlparse==0.4.1
whitenoise==5.3.0
wincertstore==0.2
